Move auth code generating utils to utils to resolve circular dependency
