+++ src/core/authentication.ts
@@ -1,5 +1,3 @@
-import crypto from 'crypto';
-import express, { Request, Response } from 'express';
 import passport from 'passport';
 import passportLocal from 'passport-local';
 import passportTotp from 'passport-totp';
@@ -8,25 +6,13 @@ import { getRepository } from 'typeorm';
 
 import config from '@config';
 
-import { cleanEmailAddress, getNextParam, sleep } from '@core/utils';
+import { cleanEmailAddress, sleep } from '@core/utils';
+import { generatePassword, hashPassword } from '@core/utils/auth';
 
 import OptionsService from '@core/services/OptionsService';
 import MembersService from '@core/services/MembersService';
 
-import Member, { Password } from '@models/Member';
-import { PermissionType } from '@models/MemberPermission';
-
-export enum AuthenticationStatus {
-	LOGGED_IN = 1,
-	NOT_LOGGED_IN = 0,
-	NOT_MEMBER = -1,
-	NOT_ADMIN = -2,
-	REQUIRES_2FA = -3
-}
-
-interface PassportUser {
-	id: string
-}
+import Member  from '@models/Member';
 
 // Add support for local authentication in Passport.js
 passport.use( new passportLocal.Strategy( {
@@ -109,127 +95,3 @@ passport.deserializeUser( async function( data, done ) {
 		return done( null, false );
 	}
 } );
-
-// Used for generating an OTP secret for 2FA
-// returns a base32 encoded string of random bytes
-export function generateOTPSecret(): Promise<string> {
-	return new Promise(resolve => {
-		crypto.randomBytes( 16, function( ex, raw ) {
-			const secret = base32.encode( raw );
-			resolve(secret.toString().replace(/=/g, ''));
-		} );
-	});
-}
-
-export function generateCode(): string {
-	return crypto.randomBytes( 10 ).toString( 'hex' );
-}
-
-// Used to create a long salt for each individual user
-// returns a 256 byte / 512 character hex string
-export function generateSalt(): Promise<string> {
-	return new Promise(resolve => {
-		crypto.randomBytes( 256, function( ex, salt ) {
-			resolve( salt.toString( 'hex' ) );
-		} );
-	});
-}
-
-// Hashes passwords through sha512 1000 times
-// returns a 512 byte / 1024 character hex string
-export function hashPassword( password: string, salt: string, iterations: number): Promise<string> {
-	return new Promise(resolve => {
-		crypto.pbkdf2( password, salt, iterations, 512, 'sha512', function( err, hash ) {
-			resolve( hash.toString( 'hex' ) );
-		} );
-	});
-}
-
-// Utility function generates a salt and hash from a plain text password
-export async function generatePassword( password: string ): Promise<Password> {
-	const salt = await generateSalt();
-	const hash = await hashPassword(password, salt, config.iterations);
-	return {
-		salt, hash,
-		iterations: config.iterations,
-		tries: 0
-	};
-}
-
-// Checks the user is logged in and activated.
-export function loggedIn( req: Request ): AuthenticationStatus {
-	// Is the user logged in?
-	if ( req.isAuthenticated() && req.user ) {
-		// Is the user active
-		if ( ! req.user.otp.activated || ( req.user.otp.activated && req.session.method == 'totp' ) ) {
-			return AuthenticationStatus.LOGGED_IN;
-		} else {
-			return AuthenticationStatus.REQUIRES_2FA;
-		}
-	} else {
-		return AuthenticationStatus.NOT_LOGGED_IN;
-	}
-}
-
-// Checks if the user has an active specified permission
-function checkPermission( req: Request, permission: PermissionType ): boolean {
-	return req.user ? req.user.quickPermissions.indexOf( permission ) !== -1 : false;
-}
-
-// Checks if the user has an active admin or superadmin privilage
-export function canAdmin( req: Request ): AuthenticationStatus {
-	// Check user is logged in
-	const status = loggedIn( req );
-	if ( status != AuthenticationStatus.LOGGED_IN ) {
-		return status;
-	} else {
-		if ( checkPermission( req, 'superadmin' ) ) return AuthenticationStatus.LOGGED_IN;
-		if ( checkPermission( req, 'admin' ) ) return AuthenticationStatus.LOGGED_IN;
-	}
-	return AuthenticationStatus.NOT_ADMIN;
-}
-
-// Checks if the user has an active superadmin privilage
-export function canSuperAdmin( req: Request ): AuthenticationStatus {
-	// Check user is logged in
-	const status = loggedIn( req );
-	if ( status != AuthenticationStatus.LOGGED_IN ) {
-		return status;
-	} else {
-		if ( checkPermission( req, 'superadmin' ) ) return AuthenticationStatus.LOGGED_IN;
-	}
-	return AuthenticationStatus.NOT_ADMIN;
-}
-
-export function handleNotAuthed( status: AuthenticationStatus, req: Request, res: Response ): void {
-	const nextUrl = req.method === 'GET' ? getNextParam(req.originalUrl) : '';
-
-	switch ( status ) {
-	case AuthenticationStatus.REQUIRES_2FA:
-		res.redirect( '/otp' + nextUrl );
-		return;
-	default:
-		req.flash( 'error', 'login-required' );
-		res.redirect( '/login' + nextUrl );
-		return;
-	}
-}
-// Checks password meets requirements
-export function passwordRequirements( password: string ): string|true {
-	if ( ! password )
-		return 'password-err-length';
-
-	if ( password.length < 8 )
-		return 'password-err-length';
-
-	if ( password.match( /\d/g ) === null )
-		return 'password-err-number';
-
-	if ( password.match( /[A-Z]/g ) === null )
-		return 'password-err-letter-up';
-
-	if ( password.match( /[a-z]/g ) === null )
-		return 'password-err-letter-low';
-
-	return true;
-}
