+++ src/core/middleware.ts
@@ -3,9 +3,9 @@ import { NextFunction, Request, RequestHandler, Response } from 'express';
 import mongoose, { Document, DocumentDefinition, FilterQuery, Model } from 'mongoose';
 import { EntityTarget, FindOneOptions, getRepository } from 'typeorm';
 
-import * as auth from '@core/authentication';
 import ajv from '@core/ajv';
 import { wrapAsync, isInvalidType } from '@core/utils';
+import * as auth from '@core/utils/auth';
 
 import OptionsService from'@core/services/OptionsService';
